Begin implementation of identity signature routines.
In order to verify that the identities for a given domain were originated
by the domain an RSA based signature is included with the domain and
element identity.  Since the identifiers are hashes the identies are
simply encrypted with an organizational private key.

This strategy may need to be revisited depending on concerns over
disclosure of the patient identitifer to the referral entity.