Implemented RSA encryption/decryption object.
The object which had previously been developed to implement a
signature of an identity was transitioned to a general purpose
object to carry out RSA encryption and decryption.

This commit also adds a public/private key pair for testing.
This key SHOULD NOT BE used for the transfer of any sensitive
data.

The beginnings of a unit test system were implemented with a
sample program to test RSA encryption of a buffer.