Implement patient identity management object.
This commit adds support for an object which creates and manages a
patient identity.  The patient identity is a 256 bit number representing
the 'fusion' of the organizational identity with the patient's identity.

The identity unit test program was updated to include the generation
and output of a sample organizational identity and the identity of
a patient in the context of the organization.