Fix error in SHA256 digesting.
The initialization of the digest context structure was being
controlled by the flag which indicated whether or not the digest
had been computed.  Unfortunately this caused the structure to
be initialized each time additional input was provided to the
hash.  Thus the last set of data added to the hash was the
only input to the hash.

A new variable was added to the state structure of the object to
track whether or not the context was initialized.   The reset
method clears this flag to allow a new hashing operation to
be initiated.