Implemented random number generation object.
An object named RandomBuffer was implemented as an interface to
the OpenSSL random number generator.  The object also supports
updating of the random state registration file which maintains
the cryptographic state of the random number generator across
multiple invocations of the object.

The genrandom utility was updated to use the new interface.