Fix multiple hash inputs.
The SHA256_hmac object was largely copied from the SHA256 object.  As
a result it inherited the problem which was just recently fixed with
the latter object.

Object poisoning was also added so the object would be rendered
non-functional by any operational issue with itself or an object
which contact it.