Add utility for RSA keypair generation.
This commit adds a small shell script to automate the generation of
an RSA keypair and the extraction of the public component of the
pair.

The sample keypair supplied with the sources are renamed to have
an output extension o .pem which his consistent with the encoding
format and standard practice.

The RSA unit test program was updated to use the new keyname
convention.

The sample organizational identity file was updated to use the
new naming nomenclature.  The section definitions were also modified
in order to simpify the automation of token generation with a
shell script.