Add identity token management object and utility.
This commit adds an object to manage patient identity tokens.  In
physical form these are expressed as ASCII delimited identities.

The token utility was also added to exercise this object.  One of
the options is a primitive search facility which allows the NPI
database to be searched for the organization which originated an
identity token.