Begin implementation of SSL communications object.
An object named Duct is introduced which will be used as the basis
for implementing secured connections between a client and a server
application.

A utility and unit test tool named Duct_test is also included.  It
has the ability to run in both client and server modes to send
secured stream messages.

A self-signed certificate is also included for testing purposes.  It
uses the previously introducted test private key as its signing
element.  Once again this certificate should be used ONLY for
testing purposes.