Completed basic implementation of SSL duct object.
The SSL duct object is now capable of sending and transmitting messages
over a secured conduit.  The unit test utility exercises the object
by sending a series of messages over a connection followed by the
receipt of a message from the server.

Authentication of the server is also supported by allowing the
specification of a certificate file.  Setting certificate verification
forces authentication of the connection while an anonymous connection
is allowed in the face of a self-signed certificate.