Added support for an Authenticator object.
This commit provides various infra-structure components needed to
implement an object for managing authentication messages.  These
messages are sent to the user identity and device identity broker
servers.  A pair of these authentication messages implement an
identity query.

An object for carrying out 256 bit AES encryptin in CBC mode was
added in order to support the Authenticator object.