Begin implementation of query client and root referral server.
The root referral server is the most fundamental element of the
identity brokerage infra-structure.  It is responsible for fielding
the identity query, authenticating the user and device and forwarding
the target identitie(s) to the brokerage servers which resolve
the holders of the identities.

The query client implements the formation of the identity query request
by encrypting the target identities in keys which can only be decrypted
by the user and device identity brokerage servers.