Complete basic implementation of authentication framework.
Skeletons of the device and user broker servers are implemented.  The
servers simply print the decode identity token sent to them by the
referral root broker.

The root referral server was updated to implement connections to these
subordinate servers.

The Makefile was updated to implement complete and correct dependencies
for the primary source files.