Actually add the modified contents.
I had failed to add modifications to the working files to the cache.
This commit addresss this problem.

The Makefile contains updated dependencies for all source files.
The root referral server is integrated for calls to the authentication
brokerage servers.  The query client is updated for better logging
of interactions.