Changes to random key generator and identity generator.
The random key generator was updated to output only the random
key unless the -F switch was specified.  The -b option was also
added to specify the number of bits which should be generated.

The identity generator was updated to output only the domain
identity and/or the user identity.   This was done to support
the generation of sample identity input files for the brokerages.