Completed implementation of authentication brokers.
The changes previously implemented in the device broker server were
integrated into the user broker server.

The query client was updated to load the three tokens necessary to
initiate a query in three separate tokens.  The code was updated to
properly load the patient elements into the two authenticators.

The root referral server and the two identity brokers are now
cooperating to authenticate a device and user.