Added and integrated identity brokerage server.
This commit adds and implements the identity brokerage server.
This server is responsible for iterating through the
authenticated identity elements and determining if this site
hosts a originating identity for the pair.

This is an initial implementation framework and does not support
the actual search for the match or return of the results.