Completed end to end identity referral.
The root referral server was updated to receive and return the
identity referrals from the brokerages.  The query client was
updated to wait and output the result of the identity query.

Query time on a 600Mhz laptop (with JB weld on the corners) was
approximately 1.6 seconds across 3,012,966 identity originating
organizations.