Implemented multiple identity token support.
The query client was updated to read all identity tokens which are
contained in a file.  The complete list of organizational identities
are composed into the two authenticators which form the identity
query.

The identity broker server was updated to handle authentication
replies which have multiple organizational identities imbedded in
them.   The broker server now prints out the list of identities
which an originating identity is located for.

The size in bytes of identity is now set as a define in the
NAAAIM.h header file.  This allows all components of the system
to use this standardized definition.