Cleaned up SSL connection shutdown handling.
The identity fabric servers were displaying messages indicating errors
were being experienced when the SSL_shutdown routine was being
called.  I traced this to always being the second invocation of
the function in the bi-directional shutdown strategy being used.

The fabric servers were updated to use a standard system for
shutting down the SSL connections.  In the process I located a
bug in the identity broker server which was causing a silent
segmentation fault.

The connections no close cleanly with no indication of error
messages.