Fixed SSL connection memory leak.
Basic stress testing of the identity fabric servers revealed a
significant memory leak.  This was caused by initializing a new
SSL connection object on every accept of a connection.

This was corrected by calling SSL_free on the connection object
in the reset method when a server object type was detected.

The fabric brokers are now able to withstand continous connection
from multiple clients with relatively constant memory
consumption.