Implemented query reply object.
This commit begins the implementation of an object which encapsulates
and manages replies from the identity brokerage servers.  There
are two levels of encoding in this object.  An inner level which
holds the contents of the reply wrapped in an outer encoding
which specifies the type of the inner encoding.

The query client, root referral server and the identity broker
were updated to use this infra-structure.  At this commit level a
null query and reply are functionally implemented.