Implemented IP referral reply.
The IDqueryReply object was updated to encode a specification
that the client should contact an identity origination server for
additional information about the client identity.

The identity broker was updated to randomly return various IP
address referrals.

Modifications were begun on the query client to support
processing of the different types of query replies.

Bugs were addressed in the root referral server and the query
client which were causing segmentation faults upon completion of
a query.