Updated and added identity generation tools.
A complete NPI search database was needed in order to begin
testing on hardward supporting the new identity broker servers.
A shell script was implemented to query the keys database on a
Postgres server and to generate all identities in the database.
This proved to be a prohibitevly slow process.

A C program was developed to read an input stream and generate
the needed identities.  This yielded a significantly faster
process.

The genid program was updated in the process of this work to take
identity precursor information on the command-line rather than
through an ASCII configuration file.