Implement Postgres connection management object.
The DBduct object was created to manage connections to a Postgres
database server.  The object implements an exec and query method
for issueing commands against the database server.  Automatic
cleanup of the result object is implemented for queries and
commands.

A test program called DBduct_test was also implemented to
exercise functionality of this object.