Updated Makefile and added result release to DBduct destructor.
The Makefile was updated to eliminate the need to have all binaries
linked against the Postgres libraries.

The destructo for the DBduct object was updated to determine whether
or not an active result structure was available.  If there was an
active structure it is freed to prevent leakage of the memory
allocated for the result.