Implemented support for database driven responses.
The identity broker now queries the database server to get
information on what type of reply is to be composed and returned
to the caller.

At this point the broker determines whether or not there is
information in the database on how to compose the reply.
Additional support needs to be added to customize the response.