Completed basic implementation of database driven queries.
The identity-broker server was updated to return three types of
identity referral information.   There are IP address referrals,
telephone number referrals and general information referrals.
The client was updated to accept these three types of referrals.

Updates were made to the IDqueryReply object were implement to
support these changes.  In addition it adds a reset() method to
enable a query reply to be re-used multiple times.