Updated configuration support for query client.
In order to support use of the query client from a graphical user
interface the -i command-line switch was added to allow the
specification of a patient identity token file from the
command-line.

The -d and -u switches were added to specify the device and user
identity token files.  These token files may also be specified in
the configuration file.