Began implementation of dynamic query display.
Investigation into the SwingWorker class provided the technology
needed to implement dynamic display of output from the query
client.  In this commit a basic implementation of this technology
was begun.

The query and log output are not currently being isolated.  The
hypothesis is that an additional worker thread needs to be
spawned in order to provide a separate thread for the log output.