Initial implementation of SMS referral.
An additional reply type was added to the IDqueryReply object.
This reply type encodes an e-mail address to be used for sending
an SMS text message to a mobile device.

The identity brokerage server was updated to encode an SMS reply
if the reply type in the repdefn table is set to 3.

The query client interprets the SMS reply message but popen'ing a
pipe to the mail program with the mail destination set to the
address returned by the brokerage server.