Initial implementation of identity provider query server.
This commit adds a basic implementation of the last major server
type required.  This server is run by a provider organization and
is responsible for providing clinical information on the subject
of a query referral.

The client currently returns a brief clinical profile containing
known allergies, medications and chronic conditions.

The query client was updated to initiate the remote query and to
print out the results.