Implemented bimodal SMS referral.
This commit implements a bimodal SMS referral.  In this model the
query client encodes an SMS referral from the professional using
the query device to the healthcare provider.

An IP referral is then sent to the identity provider.  The identity
provider server then sends an SMS message to the provider with
an authentication code and a patient identifier.

When the provider places a phone call to the query client user they
ask the professional for the authentication code displayed on
the query device.  This provides an authentication mechanism to
know they are discussing the same patient.

The query client correctly processes this type of request but there
is no support in the identity provider for the SMS referral.  The
authentication code also needs to be integrated into the system.