Implemented provider query protocol object and bimodal SMS reply.
The ProviderQuery object was created which is responsible for
managing the on the wire potocol between the query client and the
provider query server.

This object was then used to implement full bimodal support for
SMS messaging.  The provider query server looks up the referenced
patient and sends a counter SMS message to the provider with
the patient name and the verifier code to be used in authenticating
the health professional which initiated the query.

Additional work needs to be completed on the bimodal SMS messaging.
The most significant detail is the creation and deletion of
temporary files to configure the return address on the mailer.

The ProviderQuery object needs to be updated to encode a reply
to the client.