Implement multi-processor search.
The OrgSearch object was modified to allow multiple processors to
be used in order to cut search time.  On the 3 million + NPI
reference test the search time for five identities is cut from 58
seconds to approximately 10 seconds.

Multi-processor search is implemented by calling a setup function
with the number of processors which are to be used.  The search()
detects if this setup function has been called and dispatches
separate processes to to carry out the search.

The /opt/NAAAIM/etc/identity-broker.conf file was updated to to
include an additional configuration variable named processors
which sets the number of search slots to be used.