Added enhanced error checking to Duct object.
Issues were noted in testing that there spontaneous connection
failures.  The larger percentage of them were when the client
attempted to connect to the root referral server.  Other failures
were when the root referral server attempted connections to the
other brokerage servers.

The error paths in the SSL communications object were extensively
reworked.  Most error conditions now log an error message
identifying the procedure which initiated the error.

Subsequent testing with this code in place has not generated
additional errors.  Continued testing will be carried out to
determine if there are issues.