Initial framework for imbedded query client.
This set of changes establishes the framework for a client which
is suitable for inclusion on a single purpose dedicated device.
The target device has an LCD display as well as front panel keys
which can be actuated.

The LCDriver object implements the interface between the
application and the hardware I/O devices.