Implement basic support for smartcards.
This update provides the infrastructure needed to carry out basic
smartcard operations.  Included is an object which allows basic
operations such as reader and card detection.

The server which implements functionality for the imbedded client
was updated to use this infrastructure.  The client will now
detect a smartcard insertion and implement a standard query based
on insertion of the card.