Updates to 256 bit AES encryption object.
The encryption loop was modified to run the primary encryption loop
over the number of computed blocks in the incoming Buffer to be
encrypted.  The residual count is then passed to the final encryption
step.

The auto keyword was removed from all local variables as part of a
generic code modification which will remove the use of this keyword
due to the conversion of the keyword to an alternate meaning in newer
versions of the C/C++ standard.