Implement boot supervisor.
This commit adds support for the initramfs supervisor.  This
application is designed to run as pid 1 in the context of the
initial ramdisk (initramfs).

The supervisor starts a temporary instance of the TPM daemon and
unseals the root filesystem initialization vector and key.  It
configures an HPD based block device and loads the decrypted
image into the block device.  It then switches the execution root
to the block device and execs /sbin/init in the filesystem image.

The switch_root functionality was removed was moved from
load-image to the boot supervisor since this functionality is the
province of the supervisor rather then the image loader.