Import key scheduler.
The authentication system in POSSUM is based on the generation of a
one time identifier predicated on the identity of the device
initiating the connection.  This commit imports the initial and
reference implementation of the key scheduler.

The reference implementation reflected in this commit was from the One
Time Identification (OTI) pre-authentication system implemented for
Kerberos.