Initial implementation of OTEDKS.
This commit integrates the OTEDKS object into the build system.  A
test hardness (OTEDKS_test) is included to exercise the module.

The module currently works in a manner similar to its original
implementation.  At present the scheduler only fails on two of the
NIST non-overlapping tests.  The probability scores on the remaining
are 1.0.  It is unclear whether this is an issue since output from the
Linux urandom device also seems to have exceptions with a small number
of non-overlapping tests.