Fix handling of partial block encryption.
The 256bit AES encryption routine was not properly handling an
encryption payload which was not a multiple of the block size.
It appears as if this never worked but all of the code using this
routine up to this point must have been using integral block
sizes.

This problem was manifest by the decryption routine not
outputting the full encryption payload

The fix was to ensure that the residual bytes beyond a full block
were added with the encryption update routine with the standard
call to the encryption finalizer.

The AES256_cbc module was tested with both full and partial block
counts with this change in place.