Add support for token key extraction and hashed identity.
The changes to the identity token handling are in preparation for
introduction of the PossumPacket module which implements exchange
of the POSSUM protocol messages.

Since POSSUM uses One Time Epoch Differential Key Scheduling
(OTEDKS) the PossumPacket module needs access to the identity
key.  An enumerated type is added to allow the ->get_element
method to return the key.

The host side of a POSSUM exchange only needs the hashed value of
the clients identity.  This commit allows an identity size of 32
to be returned along with 256 which is the size of a standard
identity.