Add initial support for POSSUM protocol exchanges.
The primary focus of this update is the initial implementation of
an object to containerize the packets which implement the POSSUM
protocol exchanges.  This update provides an initial
implementation of a type 1 packet which is sent by the client
side of a connection to initiate the identification and
authentication exchange.

The possum binary was updated so that a transmission and
decryption of an exchange is now supported.  The current type 1
packet does not support the exchange of a public key nor does the
encryption of the authenticator packet support hardware or
software attestation.