Initial implementation of Curve25519 object.
This commit provides the framework for an initial implementation
of an object which can carry out elliptic curve key exchange
based on Daniel Bernstein's curve25519.

A test harness is included which attempts to implement a key
exchange.  At the point of this commit the procedure in the test
harness is not correctly implemented.