POSSUM packet checkpoint commit.
This is another checkpoint commit to preserve a state of functionality
with respect to POSSUM packet exchange.

The authenticator is now encoded in an ASN1 structure from which
the elements are passed.  The identity assertion and authentication
time are transmitted as a preamble to the authenticator which is
an encrypted block.

A trailing HMAC checksum is appended which verifies the identity
assertion and the authenticator.