Add protocol encoding.
This commit implements the changes needed for the client which is
initiating the connection to specify the IPsec protocol
parameters it is willing to implement.  This is done by encoding
the protocol parameters in a 32-bit unsigned integer in the
following format:

	(ENCRYPTION << 24) | (AUTHENTICATION << 16 ) |
		 (KEY_EXCHANGE_TYPE << 8 ) | KEY_EXCHANGE_SIZE

This encodes a possibility of 256 different types of each
protocol component.