Implement identity token conversion utility.
The POSSUM protocol requires the ability to verify whether or not
the party engaging in setup of the IPsec tunnel is holding a
valid identity token.  This utility provides a means for
converting a standard identity token into an identity verifier
token.

This is done by converting the identity in the token to a SHA256
representation of the identity.