Implement full POSSUM packet 1 and shared secret generation.
This commit is the result of a bungled attempt to fix a commit
which had a bug.  I had saved a copy of my working code so I
was able to recove but lost the incremental commit.

A general accessor method was implemented for the PossumPacket
object so various elements of the authenticator could be
recovered.  The PossumPacket_element enumerated type is used to
select the elements to be recovered.

The host_mode() now clears its received PossumPacket and generates
a type 1 packet to identify and authenticate itself to the
client.  The client_mode() function clears its transmitted packet
and waits for the reply packet from the host.

Both the client and host mode then use the information from the
imbeded authenticator to compute the shared to be used on each
side of the connection.