Implement Security Parameter Index (SPI) negotiation.
Support is added by this set of changes to allow the negotiation
between a client and host of the SPI value which is to be used.

The strategy is to use three separate arenas, each with 65535
entries, to choose SPI values from.  The client randomly chooses
values in both the 'host' and 'client' pools.  Both values are
guaranteed to be available on the client.

The host side of the connection attempts to use the proposed host
value and if it is in use checks the client pool allocation to
see if it is available.  If this allocation fails a proposed
entry is made from the reserve pool which the client will never
attempt an allocation from so is guaranteed to be available
unless previously needed by host/client host allocation failure
from a previous connection.