Implement generic Security Association parameter setup.
The arguements to IPsec->setup_sa method were hard-coded for
encryption and authentication types.  The key lengths which were
being extracted from the shared secret were also fixed.

The security parameters were elevated to the PossumPacket.h file
and the encode/code of the parameters are now done generically.
This will provide for straight addition of alternate encryption
and authentication systems.