Add configuration of counter-party personality.
The changes implemented in this update provide the framework for
changing the section of the configuration file being used to set
it to the identity of the counter-party to the POSSUM
configuration.  The section name used for the configuration
section is the SHA256 checksum of the organizational identity.

With these changes the abilit of a LIU node to connect to both
HUI and an external application server has been verified.