Add support for retrieval of local IP address and netmask.
This update uses the support which was added to the Netconfig
object to retrieve the IP address and netmask from an interface
to replace the local_ip and local_mask variables which were
required for configuration.

At this stage the possum utility requires an iface= variable
which defines the interface holding the address over which the
tunnel is to be built.  This is in addition to the interface=
variable which defines the alias interface to be used for the
secure network address.