Fix error detection and implement error code tracking.
The return value of the socket system call used to obtain a file
descriptor to implement the interface ioctl's was being
improperly tested against 1 rather then -1.  Correct this test.

Support for capturing the errno error code status from system
calls is implemented.  An accessor method is then added to
retrieve the error code.