Determine interface alias from available interfaces.
In order to further simplify possum configuration this set of
changes implements searching a specified physical interface for
an alias which is available to be used for configuring the
private network address.

This is implemented by sequentially serching from 1 to UINT32_MAX
for the first alias which is not found.