Remove internal error reporting from Netconfig object.
The net configuration object previusly reported the location of
an I/O error as output to stderr.  This was problematic since a
call to the ->get_address method is used to probe for an
available interface alias name.

The object has been previously configured to save its errno
status so this update removes all of the stderr based reporting
and leaves it to the caller to interpret the error message if
necessary.