Add support for capture of client IP address.
The Duct object in server mode previously converted the
connecting IP address into a hostname only.  The changes
integrated in this update capture the numberic IPV4 address in
the state structure of the connection.

A new method ->get_ipv4 was added to provide an accessor for this
address.

The Duct_test test program was updated to use this infrastructure
to print out the name and IPV4 address of the client.