Obtain remote IP address from client connection information.
In POSSUM host mode the remote IP address was previously
specified in the possum configuration file.  This information is
available from the client socket connection so drop the
configuration file requirement and use the information from the
incoming connection.

In client_mode the remote IP address has to be specified in order
to initiate the connection.  Pass the remote client information
into the setup_ipsec function from this configuration option.