Add support for initializing system manifest.
The Integrity Measurement Architecture (IMA) computes a
cryptographic hash of the contents of files which are owned by
root which are either read, executed or memory mapped.  This
allows detection of files which may be modified after load and
serves as one of the two components of an attestation message to
remote systems on the status of the system.

The current system measurement is maintained in the Linux kernel
as a linked list to which the file measurements are added as they
are encountered.  The obvious challenge to this is the current
system measurement is deterministic on the order in which the
files are encountered.

This patch implements a modification to the operating environment
boot loader to read a file manifest which is present on the root
filesystem of the system load.  This forces a deterministic measurement
of the system.  Thus the integrity of the system is defined at
system load time.  Any modifications to measured files will be
reflected in either the composite hash of all the files on
the system or its extension into the platform configuration
register which is carrying the software measurement extension of
the system status.