Update first of HurdLib.
Up to this point the project has been neglectful with respect to
updating references to the HurdLib superproject.  These changes
bring in modifications to the File and String objects which are
needed at this point in development.

We will move forward under a policy where any new HurdLib
functionality is reflected by its own commit.