Implement SoftwareStatus module.
The changes in this update provide support for an object which
queries the Integrity Measurement Architecture (IMA) for the
current software status of the platform.

Two available measurements are available.  One is the hash value
of the current template hashes.  These reflect the contents of
a file based on their filename.  The second measurement is the
hash of the file contents hashes.

A utility program named software-status is also supplied which
exercies the object and returns the current system measurement.