Implement alternate uid for image load.
The -u arguement was added to the load-image utility to specify
the uid which the encrypted virtual disk image is to be loaded
under.  This is to allow a non-root load of the file in order to
prevent IMA for registering the file in the integrity chain.

The source file was updated to use the INIT and WHACK code
compression macros.