Source cleanups to encrypted virtual disk utility.
An object leak was noted in the gen-image utility which remains
unresolved with this update.  The code in the utility itself
has been audited which suggests the leak may be in one of
the underlying objects used.

This commit converts the file to use the INIT/WHACK object
construction and destruction utilities in preparation for further
investigation in into the source of the leak.

The statically defined initialization vector and password were
removed from the file along with the printing of the
initialization vector and key which were obviously a security
issue.