Generate software status and load configuration filesystem.
The primary functional change in this commit is to decrypt and
mount an encrypted filesystem containing the device
configuration.  The load of the configuration information is done
without root privileges in order to avoid extending the
measurement state of the system with the measurement of the
encrypted image.

The measurement of the software state of the system is now done
but only displayed.  This will be leveraged in a subsequent
update to extend the boot measurement as a preamble to releasing
the identity from NVRAM.