Implement loading of initial system configuration.
The initial implementation of the init daemon is extended to
include the reading of the system configuration information.
This is done after dropping the root identity of the process in
order to prevent the configuration information from changing the
measurement status of the system.

The init daemon uses this configuration information to set the
eth0 network address.  The mode of the system is also set which
will be used to customize further configurations.