Add command line switch for specification of configuration file.
The previous version of possum made the assumption the
configuration file was in the current directory.  This commit
adds the ability to specify the location of the configuration
file via the -c command-line switch.

In the absence of a location specified on the command-line the
value of /etc/conf/possum.conf file will be used.