Implement configuration of internal tunnel.
This update adds a call to a function named
configure_internel_tunnel which is responsible for setting up the
internal IPsec link between the two motherboards in the chassis.

The function is currently hard coding the decision on how to run
the possum command by examining the mode variable.  This needs to
be updated to use the raw value of the mode variable as the
arguement to the possum utility.

There is also a hard coded delay on the client side of the
connection which needs to be configurable.