Implement restriction of server listen address.
The previous implementation of the Duct object used INADDR_ANY as
the socket bind address.  This resulted in applications
sequestered in an IPsec environment to listen on the native
interface in addition to the IPsec virtual interface.  The result
of this was to have the application operating in a 'split'
environment.

The ->set_server method was added to the object to restrict the
server to listen on a specific interface address.

The Duct_test test harness was modified to restrict the server
address to the value specified by the -h command-line arguement.