Implement direct measurement of hardware status in POSSUM.
Support for directly measuring the software status which will be
used to attest both host and client modes in POSSUM was added
using the SoftwareStatus object.

At this point the possum client is non-functional since possum is
reading its configuration files with root privileges which adds
them to the measurement state.  Subsequent updates will provide
an option for reducing provileges while the configuration files
are loaded.