Implement TPM management in init.
The changes embodied in this commit provide a framework for init
to start a TPM session once it receives control of execution.

The current implementation starts a software TPM for testing the
software in the context of the TPM emulator software.

This update also removes the check for success of setting up the
internal tunnel so execution drops to an interactive shell for
debugging the operating system instance.