Implement support for writing to TPM NVram.
The TPMcmd object is updated by these changes to include the
->nv_write method which provides support for writing the contents
of a Buffer object to a designated NVram area.

Support is provided for specifying a password in a Buffer object.
A function parameter is provided to select whether the contents
of the password buffer object should be hashed to yield a key or
whether the contents of the buffer is to be used directly as the
NVram password.