Implement new identity token format.
Up to this point in time the ASCII representation of an identity
was patient centric and would thus leak information as to the
entity which the identity represented.  This update changes the
ASCII representation to be of a generic form which does not
disclose the type of identity being represented.

A utility script is included to convert old format identity files
into the new format.