Implement ASN1 encoding of identities.
In order to faciliate the loading of a device identity into NVram
two additional methods were added to the IDtoken object.  The
methods are as follows:

	->encode

	->decode

These methods implement encoding of an identity which is
implemented in an IDtoken object into a binary ASN1
representation of the identity and its subsequent decoding.

A typographical error was also corrected which was preventing the
recognization of the authentication component of an identity.