General usability cleanup of tpm-cmd utility.
The tpm-cmd utility had grown organically in order to unit test
various methods in the TPMcmd object.  This limited the overall
usefullness of the utility.

This update isolates the various functionality into individual
components which are selectable by the first arguement to the
utility.  Current supported functionality is as follows:

	pcr_read
	pcr_extend
	nv_read
	nv_write
	lspcr

Which implements the functionality needed for further development
of the secured application environment.