Initial implementations of device identity management tools.
This commit adds the idmgr daemon which is run by the security
supervisor and initializes the device identity and then serves
requests to generation OTI authentication keys on behalf of the
device.

The id-set utility which is implemented is responsible for writing
the ASN encoded device identity into NVram.  The utility now
simply reduces the identity to ASN1 form and writes the identity
to NVram.  It needs to be extended to carry out the NVram area
definition as well.