Implement software TPM boot manager.
Implementing a boot manager via a software TPM requires a
different implementation then the boot manager based on a
hardware TPM.

This commit replicates the core boot manager functionality in a
second utility named sboot.  This will be the utility which will
be used when running the trusted application node under the
software simulator.