Implement extension of software register in security supervisor.
The security supervisor, init, was updated to include support for
extending PCR register 10 with the current software measurement.

In a hardware based TPM emulation register 10 is extended by the
kernel.  This support in the security supervisor is intended to
provide a work-alike implementation for the secure application
running under the software TPM simulator.