Load root filesystem image with reduced privileges in sboot.
This update changes the sboot utility to load the filesystem
image with reduced privileges in order to prevent the encrypted
filesystem image and the sealing file from being added to the
software measurement status.

Since the software TPM state information is saved on the
filesystem it is not possible to update the NVram contents
externally without this affecting the measurement state of the
system.

These changes allow the software TPM to conduct a measured
release of the device identity.